index.js only run an existing app
app.js returns an app which is already setup.
